requirements.txt
openpyxl>=3.1.5
python-jobspy>=1.1.82
